Add a working slurm-ssh docker image
This uses slurm instead of sge, and it can be used in the
place of "sge-ssh" for the base image of the compute images and
visualize images.

I tested it with PyFR via the [dev workflow](https://github.com/Kitware/hpccloud-services/blob/master/dev_workflow.md),
and it seemed to work fine, for both the simulation and the visualization.

The only HPCCloud web client change needed is to go into the user's
cluster settings and change the "compute" and "visualize" clusters to be
"SLURM" instead of "Sun Grid Engine".

The slurm-ssh docker image might require some additional configuration
to register the GPUs so it can work with nvidia egl, but maybe that can
be done in a separate PR.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>