Allow docker to be ran in sge images
These are the changes required to run docker in the local sge cluster images.

The changes in this PR are summarized as follows:

1. Download and install the docker client inside the sge image.
2. Add "demo" to the docker group so it has the needed permissions.
3. Mount the host docker socket inside the container.

These changes allow the "demo" user to run docker containers on the host
machine (they are "sibling" containers rather than "child" containers).

If we wish to run singularity on the demo cluster as well, we need only
to install singularity on the cluster image.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>