Replace all cases of sge with slurm
We will start using slurm by default instead of sge.

The PyFR example was tested for the simulation, osmesa visualization,
and egl visualization, and they all seem to work. We just need to
configure the clusters to be slurm, instead of sge (very soon, we will
make slurm the default cluster type instead of sge).

Note that when we use a real slurm cluster, we will probably need to
request for a GPU to run egl. But for our demo cluster, our node
already has access to the GPU, so we do not need to request it.

Also note that [this pr](https://github.com/Kitware/HPCCloud/pull/638)
is also needed to utilize all CPUs during visualization.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>